requirements.txt
openai>=1.52.0
pandas>=2.2.0
python-dotenv>=1.0.1
tenacity>=8.5.0
